requirements.txt
streamlit
torch
transformers
pillow
pytesseract
gtts
ollama
